test-requirements.txt
coverage>=3.6 # Apache-2.0
openstacksdk>=0.10.0 # Apache-2.0
oslotest>=1.10.0 # Apache-2.0
tempest>=10 # Apache-2.0
stestr>=2.0.0 # Apache-2.0
testtools>=1.4.0 # MIT
pifpaf>=0.23 # Apache-2.0
